[Refactor] #299 - 카페 API 분리
- 도메인 재분배에 따라 카페 조회 관련 API 추출
